Extract Navbar and Footer into standalone components
- Move Navbar to components/Navbar.tsx with optional animate prop
- Move Footer to components/Footer.tsx with decorative divider
- Clean up HomePage.tsx imports
