+++ apps/website/src/HomePage.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from "react";
+import { useRef } from "react";
 import {
   CanvasAsciihedron,
   useKonamiPane,
@@ -8,99 +8,13 @@ import { Button } from "./components/Button";
 import { Text } from "./components/Text";
 import { TerminalDemo } from "./components/TerminalDemo";
 import { InstallSnippet } from "./components/InstallSnippet";
-import { GitHubIcon, GitHubStarIcon, LinkedInIcon } from "./icons";
 import {
   OrchestrationContainer,
   AnimationTarget,
 } from "./components/AnimationOrchestration";
 import { AnimatedTitle } from "./components/AnimatedTitle";
-import { DISCUSSIONS_URL, RELEASES_URL, REPO_URL } from "./site";
-import { AppLink } from "./routing";
-
-function useGitHubStars(repo: string) {
-  const [stars, setStars] = useState<number | null>(null);
-
-  useEffect(() => {
-    fetch(`https://api.github.com/repos/${repo}`)
-      .then((response) => response.json())
-      .then((data) => {
-        if (typeof data.stargazers_count === "number") {
-          setStars(data.stargazers_count);
-        }
-      })
-      .catch(() => {});
-  }, [repo]);
-
-  return stars;
-}
-
-function formatStars(count: number): string {
-  if (count >= 1000) {
-    return `${(count / 1000).toFixed(1).replace(/\.0$/, "")}k`;
-  }
-
-  return String(count);
-}
-
-function Navbar() {
-  const stars = useGitHubStars("saffron-health/libretto");
-
-  return (
-    <nav
-      data-animate={AnimationTarget.Navbar}
-      style={{ opacity: 0 }}
-      className="px-8 pt-6"
-    >
-      <div className="relative mx-auto flex max-w-[800px] items-center justify-between">
-        <div className="flex items-center gap-10">
-          <AppLink href="/" className="no-underline">
-            <Text size="xl" style="serif" className="text-ink font-[200]">
-              Libretto
-            </Text>
-          </AppLink>
-          <div className="absolute left-1/2 flex -translate-x-1/2 gap-7">
-            <a
-              href={DISCUSSIONS_URL}
-              target="_blank"
-              rel="noopener noreferrer"
-              className="no-underline"
-            >
-              <Text size="sm" className="font-medium text-ink">
-                Forum
-              </Text>
-            </a>
-            <a
-              href={RELEASES_URL}
-              target="_blank"
-              rel="noopener noreferrer"
-              className="no-underline"
-            >
-              <Text size="sm" className="font-medium text-ink">
-                Changelog
-              </Text>
-            </a>
-          </div>
-        </div>
-        <div className="flex items-center gap-4">
-          <a
-            href={REPO_URL}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="flex items-center gap-1.5 text-ink/70 transition-colors hover:text-ink"
-          >
-            <GitHubStarIcon width={15} height={15} />
-            {stars !== null && (
-              <span className="text-sm font-medium">{formatStars(stars)}</span>
-            )}
-          </a>
-          <Button href="/docs/get-started/introduction" size="sm">
-            Go to docs
-          </Button>
-        </div>
-      </div>
-    </nav>
-  );
-}
+import { Navbar } from "./components/Navbar";
+import { Footer } from "./components/Footer";
 
 function Hero({
   paneUnlocked,
@@ -176,38 +90,6 @@ function Hero({
   );
 }
 
-function Footer() {
-  return (
-    <footer className="px-8 pb-8 pt-16">
-      <div className="mx-auto flex max-w-[800px] items-center justify-between">
-        <Text size="xs" className="text-muted/50">
-          © {new Date().getFullYear()} Saffron Health
-        </Text>
-        <div className="flex items-center gap-3">
-          <a
-            href="https://www.linkedin.com/company/saffron-health"
-            target="_blank"
-            rel="noopener noreferrer"
-            aria-label="Saffron Health on LinkedIn"
-            className="text-muted/50 transition-colors hover:text-muted"
-          >
-            <LinkedInIcon width={14} height={14} />
-          </a>
-          <a
-            href={REPO_URL}
-            target="_blank"
-            rel="noopener noreferrer"
-            aria-label="Libretto on GitHub"
-            className="text-muted/50 transition-colors hover:text-muted"
-          >
-            <GitHubIcon width={14} height={14} />
-          </a>
-        </div>
-      </div>
-    </footer>
-  );
-}
-
 export function HomePage() {
   const { konamiProgress, konamiCompleted, paneUnlocked, closePane } =
     useKonamiPane();
@@ -217,7 +99,7 @@ export function HomePage() {
       {!paneUnlocked && (
         <KonamiOverlay progress={konamiProgress} completed={konamiCompleted} />
       )}
-      <Navbar />
+      <Navbar animate />
       <Hero paneUnlocked={paneUnlocked} onClosePane={closePane} />
       <Footer />
     </OrchestrationContainer>
